Fix single-unit result for `months`. Revert some newer code for older code for reliability and robustness.
Fix a bug in the last step calculation of single-unit result of months in a time range.

There were residual components needed in the last equation to calculate time range in months.

The bug would make it that results in single-unit months were always almost twice as long as the actual result when the time difference was greater than 2 years. This is a very old bug and that is unfortunate it was not noticed before.

This bug does not affect any other parts of the script!

---

Revert many instances of "case statements" used as simple glob comparison tests which creeped in the code while I was doing some experiments and thought it would run faster using "case statements". These "case instances" make the code pretty unreadable and are not necessary.

Reverted the code that chooses between the various C-code date implementations. We believe the older code is more robust and is now updated to be even more robust,.

Other than that, strived to update and improve documentation (help, man pages, and repo readme.md).