Update the TESTING SUMMARY comments in the script code.
The section TESTING SUMMARY needed to be rewritten. The example mentioned in it was wrong, and we also needed to explain better what the testing results were.

Just for the sake of practicality, we bumped version of the script to v.28, but the actual working code has received no changes with this update!